Add unit tests for PHPMD collector
Also refactors Collector#config.

This is a standard configuration that all PHPMD collector instances
will use for running the engine. To avoid calling the method and
building the same object for every collector instance, we now use a
static variable to store this object.

That way, all the instances share this object and we only have to call
the method once.

Signed-off-by: Eduardo Araújo <duduktamg@hotmail.com>