Manually create the CodeClimate temp. dir. to avoid permission issues
The CodeClimate temporary directory, that will hold the configuration
JSON generated by the tool for the collector containers, has to be
world-readable so the container can open it. To make sure of that,
manually create it, properly setting up the umaks and permissions,
instead of letting Docker do it, otherwise it will be owned by
root:root.