Refactor Cyclomatic parser to properly handle duplicates
It's possible for multiple entries with the same module name to be
generated by Radon, when they represent functions or methods declared
with the same name, but that will never conflict at runtime due to magic
done by decorators.

Properly handle that case by appending the line number to the module
name whenever duplicates are found.