Handle duplicate class names in parser
Append the line number to the module name of classes declared multiple
times in the same file, as for methods and functions. This might seem
strange, but it's entirely possible if the declarations are conditional.

Update the tests and the Radon results factory accordingly to properly
verify all the weird cases.

Signed-off-by: Rafael Reggiani Manzo <rr.manzo@protonmail.com>